refactor: change structure of the project and minor fix layout
