+++ app/layout.tsx
@@ -1,15 +1,17 @@
 import type { Metadata } from 'next'
 
+import { Header } from '@/shared/ui/header'
+
+import '@/shared/styles/index.scss'
 import '@fontsource/inter/300.css'
 import '@fontsource/inter/400.css'
 import '@fontsource/inter/500.css'
 import '@fontsource/inter/600.css'
 import '@fontsource/inter/700.css'
-import '@/shared/styles/index.scss'
 
 export const metadata: Metadata = {
-  description: 'Generated by create next app',
-  title: 'Create Next App',
+  description: 'Inctagram app',
+  title: 'Inctagram',
 }
 
 export default function RootLayout({
@@ -19,7 +21,10 @@ export default function RootLayout({
 }>) {
   return (
     <html lang={'en'}>
-      <body>{children}</body>
+      <body>
+        <Header count={0} isAuth={false} />
+        <main>{children}</main>
+      </body>
     </html>
   )
 }
