+++ app/page.tsx
@@ -1,0 +1,1 @@
+export { default } from '@/pages/page'
